Refactor Pascal's Triangle implementation and add new solution for retrieving a specific row

+++ solutions/cpp/pascals_triangle.cpp
@@ -1,4 +1,3 @@
-#include <iostream>
 #include <vector>
 
 class Solution {
@@ -8,13 +7,13 @@ public:
         result.reserve(numRows);
         result.push_back({1});
 
-        for (size_t i = 1; i < numRows; ++i) {
+        for (int i = 1; i < numRows; ++i) {
             const auto& prev_row = result.back();
             std::vector<int> new_row;
             new_row.reserve(i + 1);
             new_row.push_back(1);
             
-            for (size_t j = 1; j < i; ++j) {
+            for (int j = 1; j < i; ++j) {
                 new_row.push_back(prev_row[j-1] + prev_row[j]);
             }
